--- a/pkg/policy/policy.go
+++ b/pkg/policy/policy.go
@@ -8,6 +8,7 @@ import (
 	"io/fs"
 	"path"
 	"path/filepath"
+	"slices"
 	"strings"
 
 	"github.com/go-logr/logr"
@@ -123,6 +124,7 @@ func (p *Policies) Hash(kind string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("failed to load built-in / external policies: %s: %w", kind, err)
 	}
+	slices.Sort(policies)
 	return kube.ComputeHash(policies), nil
 }