--- a/pkg/policy/policy.go
+++ b/pkg/policy/policy.go
@@ -8,7 +8,7 @@ import (
 	"io/fs"
 	"path"
 	"path/filepath"
-	"slices"
+	"sort"
 	"strings"
 
 	"github.com/go-logr/logr"
@@ -124,7 +124,9 @@ func (p *Policies) Hash(kind string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("failed to load built-in / external policies: %s: %w", kind, err)
 	}
-	slices.Sort(policies)
+	if len(policies) == 0 {
+		return "", errors.New("no policies found")
+	}
 	return kube.ComputeHash(policies), nil
 }
 
@@ -184,6 +186,7 @@ func (p *Policies) loadPolicies(kind string) ([]string, error) {
 	if len(externalPolicies) > 0 {
 		policies = append(policies, externalPolicies...)
 	}
+	sort.Strings(policies)
 	return policies, nil
 }