--- a/pkg/policy/policy.go
+++ b/pkg/policy/policy.go
@@ -124,9 +124,11 @@ func (p *Policies) Hash(kind string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("failed to load built-in / external policies: %s: %w", kind, err)
 	}
-	if len(policies) == 0 {
-		return "", errors.New("no policies found")
-	}
+	/*
+		if len(policies) == 0 {
+			return "", errors.New("no policies found")
+		}
+	*/
 	return kube.ComputeHash(policies), nil
 }