--- a/pkg/policy/policy_test.go
+++ b/pkg/policy/policy_test.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/aquasecurity/trivy-operator/pkg/kube"
 	"github.com/bluele/gcache"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -44,6 +45,53 @@ var (
 		},
 	}
 )
+
+func TestPolicies_Hash(t *testing.T) {
+	t.Run("Should return hash for valid policies", func(t *testing.T) {
+		g := NewGomegaWithT(t)
+		config := policy.NewPolicies(map[string]string{
+			"policy.valid.rego":  "<REGO_CONTENT>",
+			"policy.valid.kinds": "Pod",
+		}, testConfig{}, ctrl.Log.WithName("policy logger"), policy.NewPolicyLoader("", gcache.New(1).LRU().Build(), types.RegistryOptions{}), "1.27.1")
+
+		hash, err := config.Hash("Pod")
+		g.Expect(err).ToNot(HaveOccurred())
+		g.Expect(hash).ToNot(BeEmpty())
+	})
+
+	t.Run("Should return error when no policies are found", func(t *testing.T) {
+		g := NewGomegaWithT(t)
+		config := policy.NewPolicies(map[string]string{}, testConfig{}, ctrl.Log.WithName("policy logger"), policy.NewPolicyLoader("", gcache.New(1).LRU().Build(), types.RegistryOptions{}), "1.27.1")
+
+		_, err := config.Hash("Pod")
+		g.Expect(err).To(HaveOccurred())
+		g.Expect(err.Error()).To(ContainSubstring("no policies found"))
+	})
+
+	t.Run("Should compute correct hash for given policies", func(t *testing.T) {
+		g := NewGomegaWithT(t)
+		policies := map[string]string{
+			"policy.policy1.rego":  "package test\nallow = true",
+			"policy.policy1.kinds": "Pod",
+			"policy.policy2.rego":  "package test\nallow = false",
+			"policy.policy2.kinds": "Workload",
+		}
+		config := policy.NewPolicies(policies, testConfig{}, ctrl.Log.WithName("policy logger"), policy.NewPolicyLoader("", gcache.New(1).LRU().Build(), types.RegistryOptions{}), "1.27.1")
+
+		// Sort policy content before computing the expected hash
+		policyContents := []string{
+			"package test\nallow = true",
+			"package test\nallow = false",
+		}
+		sort.Strings(policyContents)
+
+		expectedHash := kube.ComputeHash(policyContents)
+
+		hash, err := config.Hash("Pod")
+		g.Expect(err).ToNot(HaveOccurred())
+		g.Expect(hash).To(Equal(expectedHash))
+	})
+}
 
 func TestPolicies_PoliciesByKind(t *testing.T) {
 	t.Run("Should return error when kinds are not defined for policy", func(t *testing.T) {