@simar7 2025-04-18
@afdesk in my benchmarks measuring the performance of adding sorting, there isn't any noticeable hit.

```diff
  benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/aquasecurity/trivy-operator/pkg/policy/bench
cpu: Apple M1 Pro
                 │   old.txt   │            new.txt             │
                 │   sec/op    │   sec/op     vs base           │
Policies_Hash-10   884.6n ± 1%   906.5n ± 3%  ~ (p=0.150 n=100)

                 │  old.txt   │             new.txt             │
                 │    B/op    │    B/op     vs base             │
Policies_Hash-10   673.0 ± 0%   673.0 ± 0%  ~ (p=1.000 n=100) ¹
¹ all samples are equal

                 │  old.txt   │             new.txt             │
                 │ allocs/op  │ allocs/op   vs base             │
Policies_Hash-10   18.00 ± 0%   18.00 ± 0%  ~ (p=1.000 n=100) ¹
¹ all samples are equal

```