@simar7 2025-04-18
Here's the benchmark: https://gist.github.com/simar7/256e4473f6f868169297c3169c59ead3